Refactor tests to eliminate unnecessary cloning and improve assertions
- Updated various test cases to directly assign values instead of using `.clone()`, enhancing clarity and performance.
- Adjusted assertions to ensure correctness, particularly in cases involving independent modifications.
- Replaced `vec![]` with array syntax `[]` in multiple instances for consistency and improved readability.
- Enhanced error handling assertions to prevent potential panics and ensure robust test coverage.
- Cleaned up comments and unnecessary imports for better code maintainability.

+++ glr-core/tests/gss_stack_comprehensive.rs
@@ -2,7 +2,7 @@
 
 //! Comprehensive tests for the Graph-Structured Stack (GSS) in adze-glr-core.
 
-use adze_glr_core::gss_arena::{ArenaGSS, ArenaGSSManager, ArenaGSSStats, ArenaStackNode};
+use adze_glr_core::gss_arena::{ArenaGSS, ArenaGSSManager, ArenaGSSStats};
 use adze_ir::{StateId, SymbolId};
 use typed_arena::Arena;
 
@@ -379,7 +379,7 @@ fn large_push_pop_cycle() {
     let mut gss = ArenaGSS::new(&arena, StateId(0));
     let count = 500;
     for i in 1..=count {
-        gss.push(0, StateId(i), Some(SymbolId(i as u16)));
+        gss.push(0, StateId(i), Some(SymbolId(i)));
     }
     assert_eq!(gss.active_heads[0].depth, count as usize);
     let popped = gss.pop(0, count as usize).unwrap();
